chore: centralize application and expense types
