+++ types/application.ts
@@ -1,0 +1,6 @@
+export interface ApplicationRow {
+  id: string;
+  applicant: string;
+  property: string;
+  status: string;
+}
