+++ types/expense.ts
@@ -1,0 +1,10 @@
+export interface ExpenseRow {
+  id: string;
+  date: string;
+  category: string;
+  vendor: string;
+  amount: number;
+  gst: number;
+  notes?: string;
+  receiptUrl?: string;
+}
